requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Django==3.2.5
django-filter==21.1
django-simpleui==2022.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
install==1.3.5
Markdown==3.3.6
pip==21.2.4
psycopg2==2.8.6
pycparser==2.21
pytz==2021.3
requests==2.27.1
setuptools==58.0.4
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.8
wheel==0.37.1
zipp==3.7.0
